Fedorin Danil/task2-3/src/requirements.txt
fastapi==0.54.1
uvicorn==0.11.3
pydantic
mysql-connector
